Ferdig med å gjøre bossene + fiendene litt vanskeligere med tiden. Scaler også HP bar riktig. AGINC-83  AGINC-84 AGINC-10.

+++ Vietnamese_Space_Program/Include/Entities/enemy.h
@@ -17,7 +17,7 @@ private:
     int health;
     int randomNumber;
     EntityManager* manager;
-    int mode;
+    float mode;
     int theme;
 
     // avstanden x og y fra enemy til skipet.
